requirements.txt
pandas-gbq>=0.26.1
google-cloud-bigquery>=2.34.0
google-auth>=2.6.0
pyarrow>=10.0.0
ipykernel>=7.2.0
kagglehub[pandas-datasets]
python-dotenv>=1.2.2